+++ test/src/widgets/reactive_form_field_test.dart
@@ -20,7 +20,7 @@ void main() {
     );
 
     testWidgets(
-      'Not error if formControl',
+      'No error if formControl',
       (WidgetTester tester) async {
         // Given: a ReactiveFormField with formControlName in null
         final reactiveFormField = ReactiveFormField<Object, Object>(
@@ -29,10 +29,49 @@ void main() {
           builder: (_) => Container(),
         );
 
-        // Expect: an assertion error
         expect(reactiveFormField,
             isInstanceOf<ReactiveFormField<Object, Object>>());
       },
     );
+
+    group('-> ValueAccessor for non-matching generic types', () {
+      testWidgets(
+        'throws AssertionError if ValueAccessor is not set',
+        (tester) async {
+          final field = ReactiveFormField<String, int>(
+            formControlName: null,
+            formControl: FormControl(),
+            builder: (_) => Container(),
+          );
+
+          await tester.pumpWidget(field);
+
+          expect(tester.takeException(), isA<AssertionError>());
+        },
+      );
+
+      testWidgets(
+        'throws no error if ValueAccessor is set',
+        (tester) async {
+          final field = ReactiveFormField<String, int>(
+            formControlName: null,
+            valueAccessor: _TestStringIntValueAccessor(),
+            formControl: FormControl(),
+            builder: (_) => Container(),
+          );
+
+          await tester.pumpWidget(field);
+        },
+      );
+    });
   });
+}
+
+class _TestStringIntValueAccessor extends ControlValueAccessor<String, int> {
+  @override
+  int? modelToViewValue(String? modelValue) =>
+      modelValue == null ? null : int.parse(modelValue);
+
+  @override
+  String? viewToModelValue(int? viewValue) => viewValue?.toString();
 }