+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 10.0.4
+
+## Fix
+- Fix exception when focus a control that does not exist in a group
+
+## Enhances
+- Better type definition in a control value accessor wihtin a ReactiveFormField
+
 # 10.0.3
 
 ## Fix